% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.deque.r
\name{as.deque}
\alias{as.deque}
\alias{as.deque.list}
\alias{as.deque.default}
\alias{as.deque.queue}
\alias{as.deque.stack}
\title{Convert to Deque}
\usage{
as.deque(x)

\method{as.deque}{list}(x)

\method{as.deque}{default}(x)

\method{as.deque}{queue}(x)

\method{as.deque}{stack}(x)
}
\arguments{
\item{x}{An object either to be converted to the first element of a deque
(default), or the elements of a list (or columns of a dataframe)
to be set as elements of a deque.}
}
\value{
A queue, stack, or deque.
}
\description{
Convert to Deque
}
\examples{
\dontrun{
library(dequer)
d <- as.deque(lapply(1:5, identity))
d
}

}
