% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.queue.r
\name{as.queue}
\alias{as.queue}
\alias{as.queue.list}
\alias{as.queue.default}
\alias{as.queue.deque}
\alias{as.queue.stack}
\title{Convert to Queue}
\usage{
as.queue(x)

\method{as.queue}{list}(x)

\method{as.queue}{default}(x)

\method{as.queue}{deque}(x)

\method{as.queue}{stack}(x)
}
\arguments{
\item{x}{An object either to be converted to the first element of a queue
(default), or the elements of a list (or columns of a dataframe)
to be set as elements of a queue.}
}
\value{
A queue object.
}
\description{
Convert to Queue
}
\examples{
\dontrun{
library(dequer)
q <- as.queue(lapply(1:5, identity))
q
}

}
