% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.stack.r
\name{as.stack}
\alias{as.stack}
\alias{as.stack.list}
\alias{as.stack.default}
\alias{as.stack.deque}
\alias{as.stack.queue}
\title{Convert to Stack}
\usage{
as.stack(x)

\method{as.stack}{list}(x)

\method{as.stack}{default}(x)

\method{as.stack}{deque}(x)

\method{as.stack}{queue}(x)
}
\arguments{
\item{x}{An object either to be converted to the first element of a stack
(default), or the elements of a list (or columns of a dataframe)
to be set as elements of a stack.}
}
\value{
A stack object.
}
\description{
Convert to Stack
}
\examples{
\dontrun{
library(dequer)
s <- as.stack(lapply(1:5, identity))
s
}

}
