% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/costructor.r
\name{combine}
\alias{combine}
\title{combine}
\usage{
combine(x1, x2)
}
\arguments{
\item{x1,x2}{Two different deques.}
}
\description{
Combine two deques into one deque.
}
\details{
Operates via side-effects; see examples for clarification on usage.
}
\examples{
library(dequer)
d <- deque()
push(d, 1)
push(d, 2)
popback(d)
print(d, output="full")
}

