% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deque.r
\name{deque}
\alias{deque}
\title{deque}
\usage{
deque()
}
\description{
An initializer for a deque.
}
\details{
A deque is a double-ended queue.  Insertion and deletion of
objects can happen at either end.
}
\examples{
\dontrun{
library(dequer)
d <- deque()
d
}

}

