% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peek.r
\name{peek}
\alias{peek}
\alias{peek.deque}
\alias{peek.stack}
\title{peek}
\usage{
peek(x, n = 1L)

\method{peek}{deque}(x, n)

\method{peek}{stack}(x, n)
}
\arguments{
\item{x}{A deque or a stack.}
}
\description{
View items from the front of a deque or a stack. Similar to \code{head()}.
}
\examples{
library(dequer)
d <- deque()
push(d, 1)
push(d, 2)
peek(d)

}

