% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peek.r
\name{peekback}
\alias{peekback}
\alias{peekback.deque}
\alias{peekback.queue}
\title{peekback}
\usage{
peekback(x, n = 1L)

\method{peekback}{deque}(x, n)

\method{peekback}{queue}(x, n)
}
\arguments{
\item{x}{A deque or a queue.}
}
\description{
View items from the back of a deque or a queue. Similar to \code{tail()}.
}
\examples{
library(dequer)
d <- deque()
push(d, 1)
push(d, 2)
peekback(d)

}

