% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peek.r
\name{peeking}
\alias{peek}
\alias{peek.deque}
\alias{peek.stack}
\alias{peekback}
\alias{peekback.deque}
\alias{peekback.queue}
\alias{peeking}
\title{peek/peekback}
\usage{
peek(x, n = 1L)

\method{peek}{deque}(x, n)

\method{peek}{stack}(x, n)

peekback(x, n = 1L)

\method{peekback}{deque}(x, n)

\method{peekback}{queue}(x, n)
}
\arguments{
\item{x}{A deque or a stack.}

\item{n}{The number of items to view.}
}
\description{
These methods are parallel safe.
}
\details{
View items from the front/back of a deque or a stack, for \code{peek()} or
or \code{peekback()}, respectively. Similar to \code{head()} and \code{tail()}.
}
\examples{
library(dequer)
d <- deque()
push(d, 1)
push(d, 2)
peek(d)

}

