% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/costructor.r
\name{pop}
\alias{pop}
\title{pop}
\usage{
pop(x)
}
\arguments{
\item{x}{A deque.}
}
\description{
Remove items from the front of a deque.
}
\details{
Operates via side-effects; see examples for clarification on usage.
}
\examples{
library(dequer)
d <- deque()
push(d, 1)
push(d, 2)
pop(d)
print(d, output="full")
}

