% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popping.r
\name{popping}
\alias{pop}
\alias{pop.deque}
\alias{pop.stack}
\alias{popback}
\alias{popback.deque}
\alias{popback.queue}
\alias{popping}
\title{pop/popback}
\usage{
pop(x)

\method{pop}{deque}(x)

\method{pop}{stack}(x)

popback(x)

\method{popback}{deque}(x)

\method{popback}{queue}(x)
}
\arguments{
\item{x}{A deque or a stack for \code{pop()}. A deque or queue for \code{popback()}.}
}
\description{
Remove items from the front of a deque/stack for \code{pop()}; or, 
remove items from the back of a deque/stack for \code{popback()}.
}
\details{
Operates via side-effects; see examples for clarification on usage.
}
\examples{
library(dequer)
d <- deque()
pushback(d, 1)
pushback(d, 2)
pushback(d, 3)

pop(d)
print(d, output="full")

popback(d)
print(d, output="full")

}

