% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/print.r
\name{print-deque}
\alias{print-deque}
\alias{print.deque}
\title{Printing Deques}
\usage{
\method{print}{deque}(x, ..., output = "summary")
}
\arguments{
\item{x}{A deque.}

\item{...}{Unused.}

\item{output}{A character string; determines what exactly is printed.
Options are "summary", "truncated", and "full".}
}
\description{
Printing Deques
}
\details{
If \code{output=="summary"}, then just a simple representation of the n-gram
object will be printed.

If \code{output=="truncated"}, then the first 5 items will be
printed.

If \code{output=="full"} then the full list will be printed.
}

