% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.r
\name{printer}
\alias{printer}
\alias{print.deque}
\alias{print.stack}
\alias{print.queue}
\title{Printing Deques, Stacks, and Queues}
\usage{
\method{print}{deque}(x, ..., output = "summary")

\method{print}{stack}(x, ..., output = "summary")

\method{print}{queue}(x, ..., output = "summary")
}
\arguments{
\item{x}{A queue, stack, or deque.}

\item{...}{Unused.}

\item{output}{A character string; determines what exactly is printed.
Options are "summary", "truncated", and "full".}
}
\description{
Printing Deques, Stacks, and Queues
}
\details{
If \code{output=="summary"}, then just a simple representation is printed.

If \code{output=="truncated"}, then the first 5 items will be
printed.

If \code{output=="full"} then the full data structure will be printed.
}
