% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/costructor.r
\name{push}
\alias{push}
\title{push}
\usage{
push(x, data)
}
\arguments{
\item{x}{A deque.}

\item{data}{R object to insert into the deque.}
}
\description{
Add items to the front of a deque.
}
\details{
Operates via side-effects; see examples for clarification on usage.
}
\examples{
library(dequer)
d <- deque()
push(d, 1)
push(d, 2)
print(d, output="full")
}

