% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pushback.r
\name{pushback}
\alias{pushback}
\alias{pushback.deque}
\alias{pushback.queue}
\title{pushback}
\usage{
pushback(x, data)

\method{pushback}{deque}(x, data)

\method{pushback}{queue}(x, data)
}
\arguments{
\item{x}{A deque or a queue.}

\item{data}{R object to insert at the back of a deque/queue.}
}
\description{
Add items to the back of a deque or a queue.
}
\details{
Operates via side-effects; see examples for clarification on usage.
}
\examples{
library(dequer)
d <- deque()
pushback(d, 1)
pushback(d, 2)
print(d, output="full")

}

