% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pushing.r
\name{pushing}
\alias{push}
\alias{push.deque}
\alias{push.stack}
\alias{pushback}
\alias{pushback.deque}
\alias{pushback.queue}
\alias{pushing}
\title{push/pushback}
\usage{
push(x, data)

\method{push}{deque}(x, data)

\method{push}{stack}(x, data)

pushback(x, data)

\method{pushback}{deque}(x, data)

\method{pushback}{queue}(x, data)
}
\arguments{
\item{x}{A queue, stack, or deque.}

\item{data}{R object to insert at the front of the deque/stack.}
}
\value{
Returns \code{NULL}; insertion operates via side-effects.
}
\description{
Add items to the front of a stack or deque via \code{pop()}.  
Add items to the back of a queue or deque via \code{popback()}.
}
\details{
Operates via side-effects; see examples for clarification on usage.
}
\examples{
\dontrun{
library(dequer)
d <- deque()
pushback(d, 1)
pushback(d, 2)
print(d, output="full")

push(d, 3)
print(d, output="full")
}

}

