% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queue.r
\name{queue}
\alias{queue}
\title{queue}
\usage{
queue()
}
\description{
An initializer for a queue.
}
\details{
A queue is a "first in, first out" data structure.  Like a checkout
queue (line) at a store, the first one in the queue is the first
one out.
}
\examples{
library(dequer)
q <- queue()
q

}

