% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.r
\name{rev-deque}
\alias{rev-deque}
\alias{rev.deque}
\title{rev}
\usage{
\method{rev}{deque}(x)
}
\arguments{
\item{x}{A deque.}
}
\description{
rev
}
\details{
Operates via side-effects; see examples for clarification on usage.
}
\examples{
library(dequer)
d <- deque()
for (i in 1:5) push(d, i)

print(d, output="full")
rev(d)
print(d, output="full")
}

