% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rev.r
\name{revver}
\alias{rev.deque}
\alias{rev.queue}
\alias{rev.stack}
\alias{revver}
\title{rev}
\usage{
\method{rev}{deque}(x)

\method{rev}{stack}(x)

\method{rev}{queue}(x)
}
\arguments{
\item{x}{A deque.}
}
\value{
Returns \code{NULL}; insertion operates via side-effects.
}
\description{
rev
}
\details{
Operates via side-effects; see examples for clarification on usage.
}
\examples{
\dontrun{
library(dequer)
d <- deque()
for (i in 1:5) push(d, i)

print(d, output="full")
rev(d)
print(d, output="full")
}

}

