% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sep.r
\name{sep}
\alias{sep}
\title{sep}
\usage{
sep(x, k)
}
\arguments{
\item{x}{A deque.}

\item{k}{Index to split the deque at.}
}
\value{
A deque, stack, or queue (depending on the input)
}
\description{
Spliot one object (queue/stack/deque) into two of the same type. NOTE: this
function operates via side-effects AND has a return.
}
\details{
Operates via side-effects ALTHOUGH THERE IS A NON-NULL RETURN; see examples
for clarification on usage.

The split occurs after index k.  So if the input x has, say,
elements 1 to n, then after running \code{sep(x, k)}, x will
have elements 1 to k, and the return will have values
k+1, k+2, ..., n.
}
\examples{
\dontrun{
library(dequer)
d <- deque()
for (i in 1:5) push(d, i)

### Split d into 2 deques
d_last_2 <- sep(d, 3)

### d now holds the first 3 elements
print(d, output="full")
print(d_last_5, output="full")
}

}

