% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_constructors.r
\name{deque}
\alias{deque}
\title{deque}
\usage{
deque()
}
\description{
A constructor for a deque.
}
\details{
A deque is a double-ended queue.  Insertion and deletion of
objects can happen at either end.

The implementation is a head/tail doubly linked list.
}
\examples{
\dontrun{
library(dequer)
d <- deque()
d
}

}
