% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dequer-package.r
\docType{package}
\name{dequer-package}
\alias{dequer-package}
\title{dequer}
\description{
A deque (pronounced like "deck") is a "double ended queue".  This
behaves somewhat like a list in R; but unlike an R list (which is
a contiguous array of pointers), the memory of a deque isn't
necessarily contiguous.
}
\author{
Drew Schmidt \email{wrathematics AT gmail.com}
}
\keyword{Package}

