% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dequer-package.r
\docType{package}
\name{dequer-package}
\alias{dequer-package}
\title{dequer}
\description{
Stacks, queues, and deques are list-like, abstract data types. 
These are meant to be very cheap to "grow", or insert new objects into.
A typical use case involves storing data in a list in a streaming fashion,
when you do not necessarily know how may elements need to be stored.
Unlike R's lists, data the new data structures provided here is not
necessarily stored contiguously, making insertions and deletions at the
front/end of the structure much faster.  The underlying implementation here
is new and uses a doubly linked list, and thus does not rely on R's
environments or hashing.  To avoid unnecessary data copying, most operations
on these data structures are performed via side-effects.
}
\author{
Drew Schmidt \email{wrathematics AT gmail.com}
}
\keyword{Package}

