% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop.r
\name{pop}
\alias{pop}
\alias{pop.deque}
\alias{pop.stack}
\title{pop}
\usage{
pop(x)

\method{pop}{deque}(x)

\method{pop}{stack}(x)
}
\arguments{
\item{x}{A deque or a stack.}
}
\description{
Remove items from the front of a deque or a stack.
}
\details{
Operates via side-effects; see examples for clarification on usage.
}
\examples{
library(dequer)
d <- deque()
push(d, 1)
push(d, 2)
pop(d)
print(d, output="full")

}

