% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pop.r, R/popback.r, R/pushback.r
\name{pop}
\alias{pop}
\alias{pop.deque}
\alias{pop.stack}
\alias{popback.queue}
\alias{pushback.queue}
\title{pop}
\usage{
pop(x, data)

\method{pop}{deque}(x)

\method{pop}{stack}(x)

\method{popback}{queue}(x)

\method{pushback}{queue}(x, data)
}
\arguments{
\item{x}{A deque or a stack.}
}
\description{
Remove items from the front of a deque or a stack.
}
\details{
Operates via side-effects; see examples for clarification on usage.
}
\examples{
library(dequer)
d <- deque()
push(d, 1)
push(d, 2)
pop(d)
print(d, output="full")
}

