% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/popback.r
\name{popback}
\alias{popback}
\title{popback}
\usage{
popback(x, data)
}
\arguments{
\item{x}{A deque or a queue.}
}
\description{
Remove items from the back of a deque or a queue.
}
\details{
Operates via side-effects; see examples for clarification on usage.
}
\examples{
library(dequer)
d <- deque()
push(d, 1)
push(d, 2)
popback(d)
print(d, output="full")
}

