% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/push.r
\name{push}
\alias{push}
\alias{push.deque}
\alias{push.stack}
\title{push}
\usage{
push(x, data)

\method{push}{deque}(x, data)

\method{push}{stack}(x, data)
}
\arguments{
\item{x}{A deque or a stack.}

\item{data}{R object to insert at the front of the deque/stack.}
}
\description{
Add items to the front of a deque or a stack.
}
\details{
Operates via side-effects; see examples for clarification on usage.
}
\examples{
library(dequer)
d <- deque()
push(d, 1)
push(d, 2)
print(d, output="full")

}

