% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/costructor.r
\name{pushback}
\alias{pushback}
\title{pushback}
\usage{
pushback(x, data)
}
\arguments{
\item{x}{A deque.}

\item{data}{R object to insert into the deque.}
}
\description{
Add items to the back of a deque.
}
\details{
Operates via side-effects; see examples for clarification on usage.
}
\examples{
library(dequer)
d <- deque()
pushback(d, 1)
pushback(d, 2)
print(d, output="full")
}

