% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/popback.r, R/pushback.r
\name{popback.deque}
\alias{popback.deque}
\alias{pushback}
\alias{pushback.deque}
\title{pushback}
\usage{
\method{popback}{deque}(x)

pushback(x, data)

\method{pushback}{deque}(x, data)
}
\arguments{
\item{x}{A deque or a queue.}

\item{data}{R object to insert at the back of a deque/queue.}
}
\description{
Add items to the back of a deque or a queue.
}
\details{
Operates via side-effects; see examples for clarification on usage.
}
\examples{
library(dequer)
d <- deque()
pushback(d, 1)
pushback(d, 2)
print(d, output="full")
}

