% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_constructors.r
\name{queue}
\alias{queue}
\title{queue}
\usage{
queue()
}
\description{
A queue is a "first in, first out" abstract data type.  Like a checkout
queue (line) at a store, the first item in the queue is the first
one out. New items are added to the end of the queue via \code{pushback()}.
Items are removed from the queue at the front via \code{pop()}.

The implementation is a head/tail doubly linked list.
}
\details{
A constructor for a queue.
}
\examples{
\dontrun{
library(dequer)
q <- queue()
q
}

}
