% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rev.r
\name{revver}
\alias{revver}
\alias{rev.deque}
\alias{rev.stack}
\alias{rev.queue}
\title{rev}
\usage{
\method{rev}{deque}(x)

\method{rev}{stack}(x)

\method{rev}{queue}(x)
}
\arguments{
\item{x}{A queue, stack, or deque.}
}
\value{
Returns \code{NULL}; insertion operates via side-effects.
}
\description{
rev
}
\details{
Operates via side-effects; see examples for clarification on usage.
}
\examples{
\dontrun{
library(dequer)
s <- stack()
for (i in 1:5) push(d, i)

str(s)
rev(s)
str(s)
}

}
