% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deque.r
\name{sep}
\alias{sep}
\title{sep}
\usage{
sep(x, k)
}
\arguments{
\item{x}{A deque.}

\item{k}{Index to split the deque at.}
}
\description{
Split one deque into two deques.
}
\details{
Operates via side-effects; see examples for clarification on usage.

The split occurs after index k.  So if the input x has, say,
elements 1 to n, then after running \code{sep(x, k)}, x will
have elements 1 to k, and the return will have values
k+1, k+2, ..., n.
}
\examples{
\dontrun{
library(dequer)
d <- deque()
for (i in 1:5) push(d, i)

d_new <- sep(d, 3)

print(d, output="full")
print(d_new, output="full")
}

}

