% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack.r
\name{stack}
\alias{stack}
\title{stack}
\usage{
stack()
}
\description{
A stack is a "last in, first out" data structure.
}
\details{
An initializer for a stack.
}
\examples{
\dontrun{
library(dequer)
s <- stack()
s
}

}

