% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_constructors.r
\name{stack}
\alias{stack}
\title{stack}
\usage{
stack()
}
\description{
A stack is a "last in, first out" (LIFO) abstract data type. New items are
added to the front of the stack via \code{push()}. Items are removed from
the stack at the front via \code{pop()}.

The implementation is a head/tail doubly linked list.
}
\details{
A constructor for a stack.
}
\examples{
\dontrun{
library(dequer)
s <- stack()
s
}

}
